feat(kernel): 基于 cuBLAS 实现 Linear/GEMM 算子并补充单测
- 新增 include/kernel/linear.h 与 src/kernel/linear.cu
- 使用 cublasGemmStridedBatchedEx 实现 FP32/FP16 矩阵乘法
- 支持 2D 与 batched 3D 输入，以及 trans_A / trans_B 转置选项
- 在 src/kernel/CMakeLists.txt 中链接 CUDA::cublas
- 新增 test/kernel/test_linear.cu，覆盖基础 2D、单/双转置、batched 3D、大规模随机矩阵等场景
- 修复 Tensor::ndim() 返回类型为 unsigned int，避免符号混用警告
- 优化 VecType 默认模板 static_assert 的提示信息

+++ include/core/Tensor.cuh
@@ -78,9 +78,9 @@ public:
         return size_;
     }
 
-    int ndim() const noexcept
+    unsigned int ndim() const noexcept
     {
-        return static_cast<int>(shape_.size());
+        return static_cast<unsigned int>(shape_.size());
     }
 
     const std::vector<unsigned int> &shape() const noexcept
@@ -102,7 +102,7 @@ public:
     {
         std::vector<unsigned int> s(ndim());
         unsigned int stride = 1;
-        for (int i = ndim() - 1; i >= 0; --i)
+        for (int i = static_cast<int>(ndim()) - 1; i >= 0; --i)
         {
             s[i] = stride;
             stride *= shape_[i];
